+++ src/components/Template/Template.test.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { render, screen } from "@testing-library/react";
 import { axe } from "jest-axe";
 import renderer from "react-test-renderer";
@@ -9,13 +10,14 @@ import {
   TemplateMain,
   TemplateSidebar,
   TemplateBottom,
+  SidebarPlacement,
+  TemplateMainNarrow,
 } from "./Template";
 import Placeholder from "../Placeholder/Placeholder";
+import { sidebarLabel } from "./Template.stories";
 
 const breakout = <Placeholder variant="short">Breakout</Placeholder>;
-const sidebar = "left";
 const contentTop = <Placeholder>Content Top</Placeholder>;
-const contentSidebar = <Placeholder>Left Sidebar</Placeholder>;
 const contentMain = (
   <>
     <Placeholder>Main Content</Placeholder>
@@ -24,26 +26,37 @@ const contentMain = (
 );
 const contentBottom = <Placeholder variant="short">Content Bottom</Placeholder>;
 
-const templateChildren = (
+const templateComponents = (
+  sidebar: SidebarPlacement = "none",
+  useMainNarrow: boolean = false
+) => (
   <Template sidebar={sidebar}>
     <TemplateBreakout>{breakout}</TemplateBreakout>
     <TemplateTop>{contentTop}</TemplateTop>
-    <TemplateSidebar>{contentSidebar}</TemplateSidebar>
-    <TemplateMain>{contentMain}</TemplateMain>
+    {sidebar !== "none" && !useMainNarrow && (
+      <TemplateSidebar>
+        <Placeholder>{sidebarLabel(sidebar)}</Placeholder>
+      </TemplateSidebar>
+    )}
+    {useMainNarrow ? (
+      <TemplateMainNarrow>{contentMain}</TemplateMainNarrow>
+    ) : (
+      <TemplateMain>{contentMain}</TemplateMain>
+    )}
     <TemplateBottom>{contentBottom}</TemplateBottom>
   </Template>
 );
 
 describe("Template components accessibility", () => {
   it("passes axe accessibility test", async () => {
-    const { container } = render(<>{templateChildren}</>);
+    const { container } = render(templateComponents("left"));
     expect(await axe(container)).toHaveNoViolations();
   });
 });
 
 describe("Template components", () => {
-  it("renders each section", () => {
-    render(<>{templateChildren}</>);
+  it("renders each section with left sidebar", () => {
+    render(templateComponents("left"));
 
     expect(screen.getByText("Breakout")).toBeInTheDocument();
     expect(screen.getByText("Content Top")).toBeInTheDocument();
@@ -53,26 +66,76 @@ describe("Template components", () => {
     expect(screen.getByText("Content Bottom")).toBeInTheDocument();
   });
 
-  it("renders a #mainContent id in the TemplateContent component", () => {
-    const { container } = render(<>{templateChildren}</>);
+  it("renders each section with left sidebar", () => {
+    render(templateComponents("right"));
+
+    expect(screen.getByText("Breakout")).toBeInTheDocument();
+    expect(screen.getByText("Content Top")).toBeInTheDocument();
+    expect(screen.getByText("Right Sidebar")).toBeInTheDocument();
+    expect(screen.getByText("Main Content")).toBeInTheDocument();
+    expect(screen.getByText("More Content")).toBeInTheDocument();
+    expect(screen.getByText("Content Bottom")).toBeInTheDocument();
+  });
+
+  it("renders a #mainContent id when using TemplateMain", () => {
+    const { container } = render(templateComponents("left"));
 
     expect(container.querySelector("#mainContent")).toBeInTheDocument();
     expect(screen.getByRole("main")).toHaveAttribute("id", "mainContent");
   });
 
+  it("renders a #mainContent id when using TemplateMainNarrow", () => {
+    const { container } = render(templateComponents("none", true));
+
+    expect(container.querySelector("#mainContent")).toBeInTheDocument();
+    expect(screen.getByRole("main")).toHaveAttribute("id", "mainContent");
+  });
+
+  it("passes a ref to the div wrapper element", () => {
+    const ref = React.createRef<HTMLDivElement>();
+    const { container } = render(
+      <Template ref={ref}>
+        <TemplateMain>{contentMain}</TemplateMain>
+      </Template>
+    );
+
+    expect(container.querySelectorAll("div")[0]).toBe(ref.current);
+  });
+
   it("renders the UI snapshot correctly", () => {
-    const templateComponents = renderer
-      .create(
-        <Template sidebar={sidebar}>
-          <TemplateBreakout>{breakout}</TemplateBreakout>
-          <TemplateTop>{contentTop}</TemplateTop>
-          <TemplateSidebar>{contentSidebar}</TemplateSidebar>
-          <TemplateMain>{contentMain}</TemplateMain>
-          <TemplateBottom>{contentBottom}</TemplateBottom>
-        </Template>
-      )
+    const templateComponentsLeftSidebar = renderer
+      .create(templateComponents("left"))
       .toJSON();
 
-    expect(templateComponents).toMatchSnapshot();
+    const templateComponentsRightSidebar = renderer
+      .create(templateComponents("right"))
+      .toJSON();
+
+    const templateComponentsNoSidebar = renderer
+      .create(templateComponents("none"))
+      .toJSON();
+
+    const templateComponentsMainNarrow = renderer
+      .create(templateComponents("none", true))
+      .toJSON();
+
+    const templateWithChakraProps = renderer.create(
+      <Template p="20px" color="ui.error.primary">
+        <TemplateMain>{contentMain}</TemplateMain>
+      </Template>
+    );
+
+    const templateWithOtherProps = renderer.create(
+      <Template data-testid="props">
+        <TemplateMain>{contentMain}</TemplateMain>
+      </Template>
+    );
+
+    expect(templateComponentsLeftSidebar).toMatchSnapshot();
+    expect(templateComponentsRightSidebar).toMatchSnapshot();
+    expect(templateComponentsNoSidebar).toMatchSnapshot();
+    expect(templateComponentsMainNarrow).toMatchSnapshot();
+    expect(templateWithChakraProps).toMatchSnapshot();
+    expect(templateWithOtherProps).toMatchSnapshot();
   });
 });