+++ src/components/Template/Template.tsx
@@ -1,11 +1,15 @@
-import { Box, ChakraComponent, useStyleConfig } from "@chakra-ui/react";
+import { Box, chakra, ChakraComponent, useStyleConfig } from "@chakra-ui/react";
+import { forwardRef } from "react";
+
+export const sidebarPlacementArray = ["none", "left", "right"] as const;
+export type SidebarPlacement = typeof sidebarPlacementArray[number];
 
 export interface TemplateProps {
   /** ID that other components can cross reference for accessibility purposes. */
   id?: string;
   /** Renders the `TemplateSidebar` component either on the left or
    * right side of the `TemplateMain` component. */
-  sidebar?: "none" | "left" | "right";
+  sidebar?: SidebarPlacement;
 }
 
 export interface TemplateMainProps {
@@ -23,25 +27,31 @@ const Template: ChakraComponent<
     React.PropsWithChildren<TemplateProps> & React.RefAttributes<HTMLDivElement>
   >,
   React.PropsWithChildren<TemplateProps>
-> = ({ children, sidebar = "none" }) => {
-  const styles = useStyleConfig("TemplateNew", { variant: sidebar });
+> = chakra(
+  forwardRef<HTMLDivElement, React.PropsWithChildren<TemplateProps>>(
+    ({ children, sidebar = "none" }, ref?) => {
+      const styles = useStyleConfig("Template", { variant: sidebar });
 
-  return <Box __css={styles}>{children}</Box>;
-};
+      return (
+        <Box __css={styles} ref={ref}>
+          {children}
+        </Box>
+      );
+    }
+  )
+);
 
 /**
  * This optional component renders its children above the main content
  * and spans edge-to-edge. It is most useful for `Breadcrumbs`, `Hero`,
  * or other banner-like components.
  */
-const TemplateBreakout: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateProps>
-) => {
-  const styles = useStyleConfig("TemplateNewBreakout", {});
+const TemplateBreakout: React.FC<React.PropsWithChildren> = ({ children }) => {
+  const styles = useStyleConfig("TemplateBreakout", {});
 
   return (
     <Box gridArea="breakout" __css={styles}>
-      {props.children}
+      {children}
     </Box>
   );
 };
@@ -51,9 +61,9 @@ const TemplateBreakout: React.FC<any> = (
  * will render below `TemplateBreakout` (if being used) and above the
  * main content and sidebar (if one exists).
  */
-const TemplateTop: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateProps>
-) => <Box gridArea="top">{props.children}</Box>;
+const TemplateTop: React.FC<React.PropsWithChildren> = ({ children }) => (
+  <Box gridArea="top">{children}</Box>
+);
 
 /**
  * This component renders an HTML `<main>` element with an id of "mainContent".
@@ -61,17 +71,14 @@ const TemplateTop: React.FC<any> = (
  * navigation link. The component should not be used in conjunction with
  * `TemplateMainNarrow`.
  */
-const TemplateMain: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateMainProps>
-) => {
-  const { children, id = "mainContent" } = props;
-
-  return (
-    <Box as="main" id={id} gridArea="main">
-      {children}
-    </Box>
-  );
-};
+const TemplateMain: React.FC<React.PropsWithChildren<TemplateMainProps>> = ({
+  children,
+  id = "mainContent",
+}) => (
+  <Box as="main" id={id} gridArea="main">
+    {children}
+  </Box>
+);
 
 /**
  * This component renders an HTML `<main>` element with an id of "mainContent".
@@ -81,18 +88,16 @@ const TemplateMain: React.FC<any> = (
  * lieu of `TemplateMain`, and should not be used in conjunction with
  * `TemplateSidebar`.
  */
-const TemplateMainNarrow: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateMainProps>
-) => {
-  const { children, id = "mainContent" } = props;
-  const styles = useStyleConfig("TemplateNewMainNarrow");
+const TemplateMainNarrow: React.FC<React.PropsWithChildren<TemplateMainProps>> =
+  ({ children, id = "mainContent" }) => {
+    const styles = useStyleConfig("TemplateMainNarrow");
 
-  return (
-    <Box as="main" id={id} gridArea="main" __css={styles}>
-      {children}
-    </Box>
-  );
-};
+    return (
+      <Box as="main" id={id} gridArea="main" __css={styles}>
+        {children}
+      </Box>
+    );
+  };
 
 /**
  * This optional component is used to render content in a sidebar column.
@@ -100,17 +105,17 @@ const TemplateMainNarrow: React.FC<any> = (
  * prop value of "left" or "right" must be passed to the `Template` wrapper
  * to render the correct CSS styles.
  */
-const TemplateSidebar: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateProps>
-) => <Box gridArea="sidebar">{props.children}</Box>;
+const TemplateSidebar: React.FC<React.PropsWithChildren> = ({ children }) => (
+  <Box gridArea="sidebar">{children}</Box>
+);
 
 /**
  * This optional component renders content at a max width of 1280px and
  * will always render below the main content and sidebar (if one exists).
  */
-const TemplateBottom: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateProps>
-) => <Box gridArea="bottom">{props.children}</Box>;
+const TemplateBottom: React.FC<React.PropsWithChildren> = ({ children }) => (
+  <Box gridArea="bottom">{children}</Box>
+);
 
 export {
   Template,